Fix bug in chunked-encoding handling that would occasionally miss the byte count in the second chunk.  This was due to the charIt iterator being set to the beginning of the buffer before the bufferIt iterator was incremented, causing the current buffer to be read twice.
Reproducing this bug is difficult, because it only triggers when the chunk byte count
crosses a buffer boundary.  In practice, it only happens once in roughly 5,000 responses.

As a side-effect of fixing iterateResponseLambda, some external increments
to the charIt iterator could be removed.  Also reset prevByte prior to each call to
iterateResponseLambda so that it doesn't carry over its value from the previous call.